fix warning

+++ range_queries/segtree.cpp
@@ -1,20 +1,24 @@
 //#include <bits/stdc++.h>
 #include <iostream>
 #include <vector>
+#include <cinttypes>
+#include <inttypes.h>
+#include <stdint.h>
+
 constexpr unsigned int max_height = 400000;
-void ConsTree(std::vector<int> arr, std::vector<int> segtree, long long low,
-              long long high, long long pos) {
+void ConsTree(std::vector<int> arr, std::vector<int> segtree, int64_t low,
+              int64_t high, int64_t pos) {
     if (low == high) {
         segtree[pos] = arr[low];
         return;
     }
-    long long mid = (low + high) / 2;
+    int64_t mid = (low + high) / 2;
     ConsTree(arr, segtree, low, mid, 2 * pos + 1);
     ConsTree(arr, segtree, mid + 1, high, 2 * pos + 2);
     segtree[pos] = segtree[2 * pos + 1] + segtree[2 * pos + 2];
 }
-long long query(std::vector<int> segtree, std::vector<int> lazy, long long qlow,
-                long long qhigh, long long low, long long high, long long pos) {
+int64_t query(std::vector<int> segtree, std::vector<int> lazy, int64_t qlow,
+                int64_t qhigh, int64_t low, int64_t high, int64_t pos) {
     if (low > high) {
         return 0;
     }
@@ -32,13 +36,13 @@ long long query(std::vector<int> segtree, std::vector<int> lazy, long long qlow,
     if (qlow <= low && qhigh >= high) {
         return segtree[pos];
     }
-    long long mid = (low + high) / 2;
+    int64_t mid = (low + high) / 2;
     return query(segtree, lazy, qlow, qhigh, low, mid, 2 * pos + 1) +
            query(segtree, lazy, qlow, qhigh, mid + 1, high, 2 * pos + 2);
 }
-void update(std::vector<int> segtree, std::vector<int> lazy, long long start,
-            long long end, long long delta, long long low, long long high,
-            long long pos) {
+void update(std::vector<int> segtree, std::vector<int> lazy, int64_t start,
+            int64_t end, int64_t delta, int64_t low, int64_t high,
+            int64_t pos) {
     if (low > high) {
         return;
     }
@@ -61,16 +65,16 @@ void update(std::vector<int> segtree, std::vector<int> lazy, long long start,
         }
         return;
     }
-    long long mid = (low + high) / 2;
+    int64_t mid = (low + high) / 2;
     update(segtree, lazy, start, end, delta, low, mid, 2 * pos + 1);
     update(segtree, lazy, start, end, delta, mid + 1, high, 2 * pos + 2);
     segtree[pos] = segtree[2 * pos + 1] + segtree[2 * pos + 2];
 }
 int main() {
     ;
-    long long n = 0, c = 0;
+    int64_t n = 0, c = 0;
     std::cin >> n >> c;
-    long long p = 0, q = 0, v = 0, choice = 0;
+    int64_t p = 0, q = 0, v = 0, choice = 0;
     std::vector<int> arr(n);
     std::vector<int> segtree(max_height);
     std::vector<int> lazy = {0};
